# series file type checking