--- a/tsdb/engine.go
+++ b/tsdb/engine.go
@@ -53,8 +53,8 @@ type Engine interface {
 	IteratorCost(measurement string, opt query.IteratorOptions) (query.IteratorCost, error)
 	WritePoints(points []models.Point) error
 
-	CreateSeriesIfNotExists(key, name []byte, tags models.Tags) error
-	CreateSeriesListIfNotExists(keys, names [][]byte, tags []models.Tags) error
+	CreateSeriesIfNotExists(key, name []byte, tags models.Tags, typ models.FieldType) error
+	CreateSeriesListIfNotExists(collection *SeriesCollection) error
 	DeleteSeriesRange(itr SeriesIterator, min, max int64) error
 	DeleteSeriesRangeWithPredicate(itr SeriesIterator, predicate func(name []byte, tags models.Tags) (int64, int64, bool)) error