--- a/tsdb/engine/tsm1/engine.go
+++ b/tsdb/engine/tsm1/engine.go
@@ -1628,7 +1628,7 @@ func (e *Engine) deleteSeriesRange(seriesKeys [][]byte, min, max int64) error {
 
 			name, tags := models.ParseKeyBytes(k)
 			sid := e.sfile.SeriesID(name, tags, buf)
-			if sid == 0 {
+			if sid.IsZero() {
 				continue
 			}
 
@@ -1677,7 +1677,7 @@ func (e *Engine) deleteSeriesRange(seriesKeys [][]byte, min, max int64) error {
 		// Remove the remaining ids from the series file as they no longer exist
 		// in any shard.
 		var err error
-		ids.ForEach(func(id uint64) {
+		ids.ForEach(func(id tsdb.SeriesID) {
 			name, tags := e.sfile.Series(id)
 			if err1 := e.sfile.DeleteSeriesID(id); err1 != nil {
 				err = err1