# Overview

This implements doing type checking per series. This PR introduces the `tsdb.SeriesID` and `tsdb.SeriesIDTyped` values that abstract the internal representation of a series id. This allowed the compiler to track down all of the places that needed to be modified, flagging the spots that needed consideration. The `ID` field is unfortunately exported on the types due to an issue with some reflect like tests (the diffing business). If this is considered to be too overbearing of a change, I wouldn't mind changing back to either typed uint64s or naked uint64s now that the hard work is done.

The next large portion of the PR is to have the code use a new abstraction: the `tsdb.SeriesCollection`. This type exists so that we can avoid passing around 4 parallel slices both in and out of many methods. It also helps with the logic of invalidating bad points, keeping track of partial write errors, and shuffling around the slices cheaply.

The portion of the complexity in the collection comes from how the series file handles doing lookups/writes of new series in parallel across its partitions. Since that's also the spot where we know what type the series id is supposed to be, invalidating the series at that moment becomes challenging. Using the fact that most writes won't have a problem, we make it cheap and allocation free in the case that there are no conflicts.

Lastly, one change I would like to draw attention to is in tsi1. When creating series in tsi1, it would first split the work into 8 subsets and pass those to its partitions. Then, each tsi1 partition would individually pass *their* subset of the data to the series file. Then, the series file would split that work against *its* 8 partitions, causing a total of 64 splits, and much ado. Instead, tsi1 now passes the entire collection to the series file *before* splitting the work for its individual partitions. Not only do I think this is better, it makes the job of the series file efficiently dropping incorrectly typed series easier since it can operate on all of the data at once, and no merging of the individual pieces is required.

## Benchmarks

Some benchmark results for inmem[1], names elided for space:

```
old time/op    new time/op    delta
  2.70ms ± 1%    2.64ms ± 1%  -2.31%          (p=0.000 n=8+7)
   314µs ± 2%     315µs ± 1%    ~             (p=0.442 n=8+8)
   314µs ± 4%     315µs ± 1%    ~             (p=0.536 n=8+7)
  2.02ms ± 1%    2.00ms ± 1%  -0.88%          (p=0.001 n=8+7)

old alloc/op   new alloc/op   delta
   477kB ± 0%     477kB ± 0%  -0.04%          (p=0.000 n=8+8)
   15.0B ± 0%     15.0B ± 0%    ~     (all samples are equal)
   15.0B ± 0%     15.0B ± 0%    ~     (all samples are equal)
   501kB ± 0%     501kB ± 0%  -0.04%          (p=0.000 n=8+8)

old allocs/op  new allocs/op  delta
    37.0 ± 0%      38.0 ± 0%  +2.70%          (p=0.000 n=8+8)
   0.00 ±NaN%     0.00 ±NaN%    ~     (all samples are equal)
   0.00 ±NaN%     0.00 ±NaN%    ~     (all samples are equal)
    23.0 ± 0%      23.0 ± 0%    ~     (all samples are equal)
```

tsi1 has a more comprehensive benchmark suite, but I had trouble running it, so I used inch to create 30 million series by running `inch -b 10000 -c 20 -t 10,10,100000,3 -m 1 -p 1 -v`:

| strategy    | vals/sec          | seconds        | ratio |
|-------------|-------------------|----------------|-------|
| no checking | 80829.1 val/sec.  | 371.2 seconds  | 1x    |
| checking    | 81896.5 val/sec.  | 366.3 seconds  | 0.98x |

which I believe can be attributed to noise.

---

[1] The inmem benchmarks were also benchmarking the initial add in some situations, causing the allocation amounts to vary based on how many times it ran, since only the first time allocated. The above benchmarks include backported changes to fix this.