@hercules-influx 2018-08-02
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/265371993/src/github.com/influxdata/influxdb/tsdb/index/tsi1/tsi1.go:516:6: type uint64Slice is unused (U1000)
/tmp/265371993/src/github.com/influxdata/influxdb/tsdb/series_file.go:471:6: type uint64Slice is unused (U1000)
/tmp/265371993/src/github.com/influxdata/influxdb/tsdb/series_id.go:68:2: const seriesIDStructSize is unused (U1000)
/tmp/265371993/src/github.com/influxdata/influxdb/tsdb/series_id.go:69:2: const seriesIDTypedStructSize is unused (U1000)
```