@zeebo 2018-08-02
Circle CI seems to be exposing this failure:

```
panic: interface conversion: tsm1.Value is tsm1.IntegerValue, not tsm1.FloatValue

goroutine 42196 [running]:
github.com/influxdata/influxdb/tsdb/engine/tsm1.(*floatAscendingCursor).peekCache(0x1b0bae70, 0x1, 0x2, 0x2, 0x24)
	/root/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/iterator.gen.go:410 +0xb1
github.com/influxdata/influxdb/tsdb/engine/tsm1.(*floatAscendingCursor).nextFloat(0x1b0bae70, 0x8d0c93c, 0x0, 0x1, 0x80556d8)
	/root/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/iterator.gen.go:441 +0x25
github.com/influxdata/influxdb/tsdb/engine/tsm1.(*floatIterator).Next(0x1acd4240, 0x87de2e0, 0x1, 0xf54b62b8)
	/root/go/src/github.com/influxdata/influxdb/tsdb/engine/tsm1/iterator.gen.go:236 +0x556
github.com/influxdata/influxdb/tsdb_test.TestShard_WritePoints_FieldConflictConcurrentQuery.func2(0x1acb4140, 0x1ad64e40)
	/root/go/src/github.com/influxdata/influxdb/tsdb/shard_test.go:627 +0xa95
created by github.com/influxdata/influxdb/tsdb_test.TestShard_WritePoints_FieldConflictConcurrentQuery
	/root/go/src/github.com/influxdata/influxdb/tsdb/shard_test.go:580 +0x3d2
FAIL	github.com/influxdata/influxdb/tsdb	8.708s
```

I have a feeling this change is causing it. Tickling it out with cpulimit.