@zeebo 2018-08-07
Alright, I cannot for the life of me get this to fail locally, so I'm going to start the arduous process of pushing commits with a bunch of temporary logging to extract more information out of CircleCI.