@zeebo 2018-08-08
The issue discovered here by CircleCI is now #10169.